+++ src/ezdxf/lldxf/repair.py
@@ -140,20 +140,3 @@ def filter_invalid_xdata_group_codes(tagger: Iterable[DXFTag]) -> Iterable[DXFTa
     for tag in tagger:
         if tag.code < 1000 or tag.code in VALID_XDATA_CODES:
             yield tag
-
-
-ACDB_ENTITY_GROUP_CODES = {8, 6, 62, 370, 48, 60, 420, 430, 440, 284, 347, 348, 380, 390}
-
-
-def fix_invalid_located_acdb_entity_group_codes(subclasses: List['Tags']) -> None:
-    if len(subclasses) < 2:
-        return
-
-    acdb_entity = subclasses[1]
-    if acdb_entity[0] != (100, 'AcDbEntity'):
-        return
-
-    noclass = subclasses[0]
-    for tag in noclass:
-        if tag.code in ACDB_ENTITY_GROUP_CODES:
-            acdb_entity.append(tag)